+++ src/js/features/content-generation/components/chat-action-buttons.tsx
@@ -1,0 +1,116 @@
+import React, { useState } from 'react';
+import { Button, Icon } from '@wordpress/components';
+import { __ } from '@wordpress/i18n';
+import { backup, check, copySmall } from '@wordpress/icons';
+import { useCopyToClipboard } from '@wordpress/compose';
+
+/**
+ * Props for the ChatActionButtons component
+ */
+export interface ChatActionButtonsProps {
+	onStartOver: () => void;
+	onInsertContent: ( content: string ) => void;
+	content: string;
+}
+
+/**
+ * ChatActionButtons component
+ *
+ * Action buttons for chat responses (Start Over, Copy, Insert)
+ *
+ * @param {ChatActionButtonsProps} props Component props
+ * @return {React.ReactElement} Action buttons for conversation
+ */
+export const ChatActionButtons: React.FC< ChatActionButtonsProps > = ( {
+	onStartOver,
+	onInsertContent,
+	content,
+} ) => {
+	const [ hasCopied, setHasCopied ] = useState< boolean >( false );
+
+	const onSuccessfullCopy = (): void => {
+		setHasCopied( true );
+		setTimeout( () => {
+			setHasCopied( false );
+		}, 1500 );
+	};
+	const copyRef = useCopyToClipboard( content, onSuccessfullCopy );
+
+	return (
+		<div
+			style={ {
+				display: 'flex',
+				justifyContent: 'flex-end',
+				gap: '2px',
+				marginTop: '0',
+				marginBottom: '8px',
+				flexWrap: 'wrap',
+			} }
+		>
+			<Button
+				variant="tertiary"
+				isDestructive
+				onClick={ onStartOver }
+				size="small"
+				icon={
+					<Icon
+						icon={ backup }
+						viewBox="0 0 24 24"
+						height={ 16 }
+						width={ 16 }
+					/>
+				}
+				iconPosition="right"
+				style={ { paddingInlineEnd: '4px' } }
+			>
+				{ __( 'Start Over', 'classifai' ) }
+			</Button>
+			<Button
+				ref={ copyRef }
+				variant="tertiary"
+				size="small"
+				icon={
+					hasCopied ? (
+						<Icon
+							icon={ check }
+							viewBox="0 0 24 24"
+							height={ 16 }
+							width={ 16 }
+						/>
+					) : (
+						<Icon
+							icon={ copySmall }
+							viewBox="0 0 24 24"
+							height={ 16 }
+							width={ 16 }
+						/>
+					)
+				}
+				disabled={ hasCopied }
+				iconPosition="right"
+				style={ { paddingInlineEnd: '4px' } }
+			>
+				{ hasCopied
+					? __( 'Copied!', 'classifai' )
+					: __( 'Copy', 'classifai' ) }
+			</Button>
+			<Button
+				variant="tertiary"
+				onClick={ () => onInsertContent( content ) }
+				size="small"
+				icon={
+					<Icon
+						icon={ check }
+						viewBox="0 0 24 24"
+						height={ 20 }
+						width={ 20 }
+					/>
+				}
+				iconPosition="right"
+				style={ { paddingInlineEnd: '4px' } }
+			>
+				{ __( 'Insert', 'classifai' ) }
+			</Button>
+		</div>
+	);
+};
