--- src/js/features/content-generation/components/index.js
@@ -1,5 +1,0 @@
-export * from './render-card-body';
-export * from './render-card-footer';
-export * from './render-card';
-export * from './render-data';
-export * from './render-error';
