+++ src/js/features/content-generation/components/types.ts
@@ -1,0 +1,4 @@
+export interface ConversationEntry {
+	prompt: string;
+	completion: string | null;
+}
