+++ src/js/features/content-generation/index.js
@@ -1,166 +1,46 @@
-import './fabian';
-/**
- * External Dependencies.
- */
-import { select } from '@wordpress/data';
-import { PluginPostStatusInfo, PostTypeSupportCheck } from '@wordpress/editor';
-import {
-	Button,
-	TextareaControl,
-	Modal,
-	Flex,
-	FlexItem,
-} from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
+import React, { useEffect } from 'react';
+import { createRoot } from '@wordpress/element';
+import domReady from '@wordpress/dom-ready';
 import { registerPlugin } from '@wordpress/plugins';
-import { useState } from '@wordpress/element';
-import apiFetch from '@wordpress/api-fetch';
+import { ChatUI } from './components/chat-ui';
 
 /**
- * Internal Dependencies.
+ * RenderChatUI component
+ *
+ * Renders the ChatUI component into the editor
+ *
+ * @return {React.ReactElement|null} Component that renders nothing directly
  */
-import { DisableFeatureButton } from '../../components';
-import { RenderData, RenderError } from './components';
-
-const ContentGenerationPlugin = () => {
-	const [ isLoading, setIsLoading ] = useState( false );
-	const [ isOpen, setOpen ] = useState( false );
-	const [ error, setError ] = useState( false );
-	const [ summary, setSummary ] = useState( '' );
-	const [ conversation, setConversation ] = useState( [] );
-	const [ iterating, setIterating ] = useState( false );
-
-	const iterate = () => {
-		setIterating( true );
-		setSummary( '' );
-	};
-
-	const openModal = () => setOpen( true );
-	const closeModal = () => {
-		setConversation( [] );
-		setSummary( '' );
-		setError( false );
-		setIterating( false );
-		setOpen( false );
-	};
-
-	const startOver = () => {
-		setConversation( [] );
-		setSummary( '' );
-		setError( false );
-	};
-
-	const postId = select( 'core/editor' ).getCurrentPostId();
-
-	const getContent = async () => {
-		const title = select( 'core/editor' ).getEditedPostAttribute( 'title' );
-
-		setIsLoading( true );
-		apiFetch( {
-			path: '/classifai/v1/create-content',
-			method: 'POST',
-			data: { id: postId, summary, title, conversation },
-		} ).then(
-			async ( res ) => {
-				setConversation( [
-					...conversation,
-					{
-						prompt: summary,
-						completion: res,
-					},
-				] );
-				setError( false );
-				setIterating( false );
-				setIsLoading( false );
-			},
-			( err ) => {
-				setError( err?.message );
-				setIsLoading( false );
-			}
+export const RenderChatUI = () => {
+	useEffect( () => {
+		const editorIframe = document.querySelector(
+			'.editor-visual-editor.is-iframed'
 		);
-	};
 
-	return (
-		<PluginPostStatusInfo className="classifai-post-status">
-			{ isOpen && (
-				<Modal
-					title={ __( 'Generate content', 'classifai' ) }
-					onRequestClose={ closeModal }
-					isFullScreen={ false }
-					size="large"
-					className="content-modal"
-				>
-					{ conversation.length < 1 && (
-						<TextareaControl
-							rows="5"
-							label={ __( 'Article summary', 'classifai' ) }
-							onChange={ ( value ) => {
-								setSummary( value );
-							} }
-							value={ summary }
-							disabled={ isLoading }
-						/>
-					) }
+		if ( ! editorIframe || ! editorIframe.parentNode ) {
+			return;
+		}
 
-					{ conversation.length < 1 && (
-						<Flex justify="flex-end">
-							<FlexItem>
-								<Button
-									variant="secondary"
-									onClick={ closeModal }
-									disabled={ isLoading }
-								>
-									{ __( 'Cancel', 'classifai' ) }
-								</Button>
-							</FlexItem>
-							<FlexItem>
-								<Button
-									variant="primary"
-									onClick={ getContent }
-									isBusy={ isLoading }
-									disabled={ isLoading || ! summary }
-								>
-									{ __( 'Submit', 'classifai' ) }
-								</Button>
-							</FlexItem>
-						</Flex>
-					) }
+		const rootElement = document.createElement( 'div' );
+		editorIframe.parentNode.appendChild( rootElement );
 
-					{ conversation.length >= 1 && (
-						<RenderData
-							data={ conversation }
-							closeModal={ closeModal }
-							startOver={ startOver }
-							iterating={ iterating }
-							iterate={ iterate }
-							setSummary={ setSummary }
-							summary={ summary }
-							isLoading={ isLoading }
-							getContent={ getContent }
-						/>
-					) }
+		const root = createRoot( rootElement );
+		root.render( <ChatUI /> );
 
-					{ ! isLoading && error && <RenderError error={ error } /> }
+		return () => {
+			root.unmount();
+			if ( editorIframe.parentNode && rootElement.parentNode ) {
+				editorIframe.parentNode.removeChild( rootElement );
+			}
+		};
+	}, [] );
 
-					{ ! isLoading && (
-						<DisableFeatureButton feature="feature_content_generation" />
-					) }
-				</Modal>
-			) }
-			<PostTypeSupportCheck supportKeys="editor">
-				<Button
-					className="content"
-					variant="secondary"
-					onClick={ () => openModal() }
-					isBusy={ isLoading }
-				>
-					{ __( 'Generate content', 'classifai' ) }
-				</Button>
-			</PostTypeSupportCheck>
-		</PluginPostStatusInfo>
-	);
+	return null;
 };
 
-registerPlugin( 'classifai-plugin-content-generation', {
-	render: ContentGenerationPlugin,
+// Initialize the plugin when the DOM is ready
+domReady( () => {
+	registerPlugin( 'classifai', {
+		render: RenderChatUI,
+	} );
 } );
